# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AccountLinkInput {
  /** The anonymous device id whose data should be carried over. */
  deviceId: string;
}

export interface AccountLinkResult {
  /** True if a device row was found and merged; false on no-op. */
  linked: boolean;
  conversationsMoved: number;
  vocabSelectionsMoved: number;
  /** The Clerk-verified primary email stored on the account. */
  email?: string | null;
}

export interface ScanRequest {
  /** Base64-encoded image data */
  imageBase64: string;
  /** Language to learn (e.g. "Spanish", "French", "Japanese") */
  targetLanguage: string;
  /** User's native language (e.g. "English") */
  nativeLanguage: string;
  /** Conversation difficulty tier ("Beginner", "Intermediate", "Advanced") */
  difficulty?: string;
}

export interface ScanResult {
  conversationId: number;
  /** Identified item name in English */
  itemName: string;
  /** Identified item name in target language */
  itemNameTranslated: string;
  /** First AI message to start the conversation */
  initialMessage: string;
}

export interface ApiError {
  error: string;
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
  lastOpenedAt?: string | null;
  /** Most recent end-of-conversation grade (0-100), if graded */
  gradeScore?: number | null;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface OpenaiConversationInput {
  title: string;
}

export interface OpenaiMessageInput {
  content: string;
  /** Current learning language (re-anchored each turn) */
  targetLanguage?: string;
  /** Conversation difficulty tier ("Beginner", "Intermediate", "Advanced") */
  difficulty?: string;
}

export interface OpenaiGradeMistake {
  error: string;
  correction: string;
}

export interface OpenaiConversationGrade {
  /** Overall score from 0 to 100 */
  score: number;
  summary: string;
  strengths: string[];
  mistakes: OpenaiGradeMistake[];
  suggestions: string[];
  gradedAt?: string | null;
}

export interface OpenaiGradeInput {
  targetLanguage?: string;
  nativeLanguage?: string;
  difficulty?: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  lastOpenedAt?: string | null;
  difficulty?: string | null;
  grade?: OpenaiConversationGrade | null;
  messages: OpenaiMessage[];
}

export interface VocabularyEntry {
  word: string;
  language: string;
  count: number;
  firstSeenAt: string;
  conversationId: number;
  conversationTitle: string;
}

export interface OpenaiError {
  error: string;
}

export interface VocabBankWord {
  word: string;
  translation: string;
  /** beginner | intermediate | advanced | expert */
  level: string;
}

export interface VocabBank {
  words: VocabBankWord[];
}

export interface VocabSearchInput {
  query: string;
  targetLanguage: string;
  nativeLanguage: string;
}

export interface VocabSearchResults {
  results: VocabBankWord[];
}

export interface VocabSelection {
  id: number;
  targetLanguage: string;
  level: string;
  word: string;
  translation: string;
  createdAt: string;
}

export interface VocabSelectionInput {
  targetLanguage: string;
  level: string;
  word: string;
  translation: string;
}

export interface VocabExampleInput {
  word: string;
  targetLanguage: string;
  nativeLanguage: string;
}

export interface VocabExample {
  /** Example sentence in the target language */
  sentence: string;
  /** Translation of the sentence in the native language */
  translation: string;
}

export interface VocabCheckInput {
  word: string;
  /** The learner's own sentence */
  sentence: string;
  targetLanguage: string;
  nativeLanguage: string;
}

export interface VocabCheck {
  correct: boolean;
  /** Encouraging feedback in the native language */
  feedback: string;
  /** A corrected version of the sentence, if needed */
  correction: string;
}

export interface SentencePhrase {
  /** Situation grouping, e.g. greetings | basics | directions */
  category: string;
  /** The phrase written in the target language */
  phrase: string;
  /** Translation of the phrase in the native language */
  translation: string;
}

export interface SentenceBank {
  sentences: SentencePhrase[];
}

export interface StartChatInput {
  targetLanguage: string;
  nativeLanguage: string;
}

export interface StartChatResult {
  conversationId: number;
  /** The tutor's opening message in the target language */
  initialMessage: string;
}

export type GetVocabBankParams = {
targetLanguage: string;
nativeLanguage: string;
};

export type ListVocabSelectionsParams = {
targetLanguage: string;
};

export type GetSentenceBankParams = {
targetLanguage: string;
nativeLanguage: string;
};