# lib/api-zod/src/generated/types/accountLinkInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AccountLinkInput {
  /** The anonymous device id whose data should be carried over. */
  deviceId: string;
}