# lib/api-zod/src/generated/types/accountLinkResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AccountLinkResult {
  /** True if a device row was found and merged; false on no-op. */
  linked: boolean;
  conversationsMoved: number;
  vocabSelectionsMoved: number;
  /** The Clerk-verified primary email stored on the account. */
  email?: string | null;
}